+++ inlang/source-code/end-to-end-tests/paraglide-js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inlang/paraglide-js-e2e
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [eb941fe]
+- Updated dependencies [9566348]
+  - @inlang/paraglide-js@1.9.0
+
 ## 0.0.23
 
 ### Patch Changes